Use integers instead of size_ts for token types & types details - rename init_env_struct - use new env method in lexer tests

+++ includes/lexer.h
@@ -6,7 +6,7 @@
 /*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/12 15:14:05 by jjaniec           #+#    #+#             */
-/*   Updated: 2018/09/28 21:26:05 by jjaniec          ###   ########.fr       */
+/*   Updated: 2018/09/30 18:56:20 by jjaniec          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -126,13 +126,13 @@ int						is_separator(char c);
 
 int						is_operator(char c);
 
-t_lexeme				*create_lexeme(size_t type, char *data, \
-							size_t type_details, char **lexeme_begin_end_ptrs);
+t_lexeme				*create_lexeme(int type, char *data, \
+							int type_details, char **lexeme_begin_end_ptrs);
 
-size_t					get_lexeme_type(char *s, int *pos, \
-							char **data, size_t *type_details);
+int						get_lexeme_type(char *s, int *pos, \
+							char **data, int *type_details);
 
-size_t					lexeme_type_word(char *s, int *pos, char **data);
+int						lexeme_type_word(char *s, int *pos, char **data);
 
 void					handle_quotes_expansions(char **data);
